requirements.txt
discord.py>=1.7.2
tweepy==3.10.0
discord-py-slash-command==2.0.2
asyncio>=3.4.3